pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["IlyaBulatau <ilyabulatau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
jinja2 = "^3.1.3"
websockets = "^12.0"
python-multipart = "^0.0.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"
requests = "^2.31.0"
aiohttp = "^3.9.5"
email-validator = "^2.1.1"
taskiq = {extras = ["reload"], version = "^0.11.9"}
taskiq-aio-pika = "^0.4.1"
pyyaml = "^6.0.2"
aioboto3 = "^13.3.0"
python-slugify = "^8.0.4"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-factoryboy = "^2.7.0"
coverage = "^7.5.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
pre-commit = "^3.7.0"
mypy = "^1.13.0"
types-aioboto3 = "^13.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
